+++ lib/src/buttons/text_only_dropdown_button.dart
@@ -102,7 +102,7 @@ class _TextOnlyDropdownButtonState
 
     return SmartTooltipText(
       text: displayText,
-      config: _config,
+      tooltipTheme: effectiveTooltipTheme,
       style: isHint ? _config.hintStyle : _config.textStyle,
       textAlign: _config.textAlign,
       maxLines: _config.maxLines,
@@ -118,7 +118,7 @@ class _TextOnlyDropdownButtonState
   Widget buildItemWidget(String item, bool isSelected) {
     return SmartTooltipText(
       text: item,
-      config: _config,
+      tooltipTheme: effectiveTooltipTheme,
       style: isSelected
           ? _config.selectedTextStyle ?? _config.textStyle
           : _config.textStyle,